+++ include/encryptsrv.h
@@ -3,7 +3,7 @@
 
 #include <stdio.h>
 
-int encsrv_verify_blsheader(int fd, void* buffer, size_t length, int type);
+int encsrv_verify_blsheader(int fd, void* buffer, unsigned long length, unsigned long type);
 
 int encsrv_decrypt_header(int fd, void* buffer, size_t length, int type);
